+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## unreleased
 ### Changed
 - Require NodeJS 14
+- Replace `xml2js` with `fast-xml-parser`
 
 ## [2.0.2] – 2022-07-14
 ### Fixed